requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
beautifulsoup4==4.8.1
blis==0.4.1
Bottleneck==1.3.1
catalogue==0.0.8
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cmake==3.15.3
colorama==0.4.1
cycler==0.10.0
cymem==2.0.3
dataclasses==0.7
decorator==4.4.1
dlib==19.8.1
face-recognition==1.2.3
face-recognition-models==0.3.0
fastai==1.0.59
fastprogress==0.1.22
Flask==1.1.1
idna==2.8
imageio==2.6.1
importlib-metadata==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
more-itertools==8.0.0
murmurhash==1.0.2
networkx==2.4
numexpr==2.7.0
numpy==1.17.4
nvidia-ml-py3==7.352.0
opencv-python==4.1.2.30
packaging==19.2
pandas==0.25.3
PassportEye==1.4.1
pdfminer3k==1.3.1
Pillow==6.2.1
plac==1.1.3
pluggy==0.13.1
ply==3.11
preshed==3.0.2
py==1.8.0
pyparsing==2.4.5
pytesseract==0.3.0
pytest==5.3.1
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.2
requests==2.22.0
scikit-image==0.16.2
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
soupsieve==1.9.5
spacy==2.2.3
srsly==0.2.0
thinc==7.3.1
torch==1.3.1
torchvision==0.4.2
tqdm==4.40.0
urllib3==1.25.7
wasabi==0.4.2
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
